AddTask Second Commit
-Сохранение задания на сервер
-Исправлены несколь ToDo
-Переименован класс информации о пользователе

+++ app/src/main/java/com/x3noku/daily_maps_android/Task.kt
@@ -6,7 +6,7 @@ class Task internal constructor() {
     var nameOfTask: String? = null
     var startTimeOfTask: Int = 0
     var durationOfTask: Long = 0
-    var priorityOfTask: Byte = 0
+    var priorityOfTask: Int = 0
     var coordinatesOfTask: LatLng? = null
 
     init {
